% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRandGeneClusterPercentage.R
\name{checkRandGeneClusterPercentage}
\alias{checkRandGeneClusterPercentage}
\title{checkRandGeneClusterPercentage}
\usage{
checkRandGeneClusterPercentage(results, i, n_rounds, alpha, k = 2)
}
\description{
Function to check that not too many random genes are clustering with the
discovered bait.  Will return a warning if that is the case.
}
