% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseBaitClustNum.R
\name{chooseBaitClustNum}
\alias{chooseBaitClustNum}
\title{chooseBaitClustNum}
\usage{
chooseBaitClustNum(cor_mat, potential_bait, max_clusters = 5)
}
\description{
Internal function to choose number of clusters in performSplit
}
