% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAvgDBIndex.R
\name{computeAvgDBIndex}
\alias{computeAvgDBIndex}
\title{computeAvgDBIndex}
\usage{
computeAvgDBIndex(bait_genes, exp_mat, k = 2, alpha = 5, n_rounds = 50)
}
\description{
Internal function to compute tightness of gene set when probing for fishability
}
