% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAvgDBIndexCosSpectral.R
\name{computeAvgDBIndexCosSpectral}
\alias{computeAvgDBIndexCosSpectral}
\title{computeAvgDBIndexCosSpectral}
\usage{
computeAvgDBIndexCosSpectral(
  bait_genes,
  exp_mat,
  method,
  k = 2,
  alpha = 5,
  n_rounds = 50
)
}
\description{
Internal function to compute tightness of gene set when probing for fishability
}
