% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAvgDBIndexCosUMAP.R
\name{computeAvgDBIndexCosUMAP}
\alias{computeAvgDBIndexCosUMAP}
\title{computeAvgDBIndexCosUMAP}
\usage{
computeAvgDBIndexCosUMAP(
  bait_genes,
  exp_mat,
  method,
  k = 2,
  alpha = 5,
  n_rounds = 50,
  n_neighbors = 15
)
}
\arguments{
\item{bait_genes}{bait or potential bait}

\item{exp_mat}{expression matrix}

\item{method}{similarity metric}

\item{k}{clusters}

\item{alpha}{number of genes to sample}

\item{n_rounds}{number of rounds of sampling}

\item{n_neighbors}{number of neighbors for UMAP}
}
\description{
Internal function to compute tightness of gene set when probing for fishability
}
