% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAvgDBIndexDist.R
\name{computeAvgDBIndexDist}
\alias{computeAvgDBIndexDist}
\title{computeAvgDBIndexDist}
\usage{
computeAvgDBIndexDist(
  bait_genes,
  exp_mat,
  method,
  k = 2,
  alpha = 5,
  n_rounds = 50
)
}
\arguments{
\item{bait_genes}{bait or potential bait}

\item{exp_mat}{expression matrix}

\item{method}{similarity metric}

\item{k}{clusters}

\item{alpha}{number of genes to sample}

\item{n_rounds}{number of rounds of sampling}
}
\description{
Internal function to compute tightness of gene set when probing for fishability
}
