% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAvgDBIndexUMAP.R
\name{computeAvgDBIndexUMAP}
\alias{computeAvgDBIndexUMAP}
\title{computeAvgDBIndexUMAP}
\usage{
computeAvgDBIndexUMAP(
  bait_genes,
  exp_mat,
  k = 2,
  alpha = 5,
  n_rounds = 50,
  n_neighbors = 15
)
}
\description{
Internal function to compute tightness of gene set when probing for fishability
}
