% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosineSimMatrix.R
\name{cosineSimMatrix}
\alias{cosineSimMatrix}
\title{cosineSimMatrix}
\usage{
cosineSimMatrix(mat)
}
\arguments{
\item{mat}{matrix to compute similarity}
}
\value{
similarity matrix between rows of \code{mat}
}
\description{
Internal function to get cosine similarity
}
