% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUIDensity.R
\name{EUIDensity}
\alias{EUIDensity}
\title{EUIDensity}
\usage{
EUIDensity(X, c0, c1, alpha0, alpha1, g0, g1, ...)
}
\arguments{
\item{X}{CFR vector}

\item{c0}{from getCutoff}

\item{c1}{from getCutoff}

\item{alpha0}{from getCutoff}

\item{alpha1}{from getCutoff}

\item{g0}{from getCutoff}

\item{g1}{from getCutoff}

\item{...}{additional arguments}
}
\description{
Helper function for \code{getCutoff}.
}
