% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneFishing.R
\name{geneFishing}
\alias{geneFishing}
\title{geneFishing}
\usage{
geneFishing(
  X,
  bait_genes,
  alpha = 5,
  fishing_rounds = 1000,
  k = 2,
  min_tightness = 0.5,
  n_probing_rounds = 100,
  umap = TRUE,
  ncores = 2,
  min_bait_genes = 5,
  bait_index = 1,
  method = c("spearman", "pearson", "kendall")
)
}
\arguments{
\item{X}{matrix of gene expression where the rows are genes and the
columns are samples (cells or individuals).}

\item{bait_genes}{genes to use as bait to fish out similar genes from the
\code{X}.  These genes should be a subset of the rownames of \code{X}. It
can also be the output of the \code{probeFishability} function.  In that case
it will by default use the tightest bait set.  To change bait set you can use
the \code{bait_index} parameter.}

\item{alpha}{controls number of random genes that are sampled in each round
of fishing.  It will sample \code{alpha} times the number of genes in
\code{bait_genes}.  The default is 5.  The stronger the bait separates from
samples of random genes the larger \code{alpha} the algorithm can handle.}

\item{fishing_rounds}{number of rounds of fishing to do.  The default is 1,000.}

\item{k}{number of clusters to use in geneFishing.  The default is 2, however
if you fished out too many genes with \code{k = 2} it can help to increase to
\code{k = 3}.}

\item{min_tightness}{parameter that controls how tight the set of genes needs
to be to be considered as bait.  Tightness is measured computing the ratio of
the median distance between all potential bait genes to the distance between the
medoid of the potential bait genes and a set of randomly sampled genes. The
average of these tightness measures are taken over \code{n_probing_rounds} random samples
of genes.}

\item{n_probing_rounds}{number of random samples to be used in assessing fishability
of a set of genes.  The default is 100.  .}

\item{umap}{indicator of whether the computation should be done using UMAP.
The default is TRUE.  When umap = FALSE it will use spectral coordinates.}

\item{ncores}{number of cores to do parallel computation on if parallel = TRUE.}

\item{min_bait_genes}{minimum number of genes for bait set to be viable.}

\item{bait_index}{index of \code{bait_sets} list from \code{probeFishability()}, to
use.  Defaults to 1, which corresponds to the tightest bait.}

\item{method}{input to \code{cor()} function telling what type of correlation
to use. Defaults to spearman.}
}
\value{
A data.frame with the capture frequency rate for all genes
}
\description{
Function to perform gene fishing on a gene expression matrix.
}
\examples{

}
