% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneFishingUMAP.R
\name{geneFishingUMAP}
\alias{geneFishingUMAP}
\title{geneFishingUMAP}
\usage{
geneFishingUMAP(
  exp_mat,
  bait_genes,
  pool_genes,
  alpha,
  fishing_rounds,
  k,
  method
)
}
\arguments{
\item{exp_mat}{matrix of gene expression where the rows are genes and the
columns are samples (cells or individuals).}

\item{bait_genes}{genes to use as bait to fish out similar genes from the
\code{exp_mat}.  These genes should be a subset of the rownames of \code{exp_mat}.}

\item{alpha}{controls number of random genes that are sampled in each round
of fishing.  It will sample \code{alpha} times the number of genes in
\code{bait_genes}.  The default is 5.  The stronger the bait separates from
samples of random genes the larger \code{alpha} the algorithm can handle.}

\item{fishing_rounds}{number of rounds of fishing to do.  The default is 1,000.}

\item{k}{number of clusters to use in geneFishing.  The default is 2, however
if you fished out too many genes with \code{k = 2} it can help to increase to
\code{k = 3}.}
}
\description{
Function to do UMAP version of geneFishing
}
