% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoff.R
\name{getCutoff}
\alias{getCutoff}
\title{getCutoff}
\usage{
getCutoff(
  X,
  bait_indices,
  unit,
  left_lb = NULL,
  left_ub = NULL,
  right_ub = NULL,
  right_lb = NULL,
  mu = NULL,
  eps_l = 0.01,
  eps_r = 0.01,
  topK = 1,
  max_pct = 0.025
)
}
\arguments{
\item{X}{CFR vector from GeneFishing results from\code{geneFishing()}.}

\item{unit}{searching unit for grid search, defaults to 1 / number of fishing
rounds}

\item{left_lb}{lower bound of left peak of distribution, defaults to minimum
of X plus one unit}

\item{left_ub}{upper bound of the left peak of distribution, defaults to
left_lb + 0.2}

\item{right_ub}{upper bound of right peak of distribution, defaults to the
maximum of X minus one unit}

\item{right_lb}{lower bound of right peak of distribution, defaults to
right_ub - 0.2}

\item{mu}{excepted flat part of distribution. defaults to the middle of the
left_ub and right_lb}

\item{eps_l}{the gap between the left part and the middle part.}

\item{eps_r}{the gap between the right part and the middle part}

\item{topK}{parameter for finding cutoff}

\item{max_pct}{pct of genes if clustered with bait 100\% we use as cutoff}
}
\value{
CFR cutoff for labeling fished genes
}
\description{
Function to get CFR cutoff from \code{geneFishing()} results. This method
assumes the model is decreasing, flat and increasing. It detects the
boundary ofmthe decreasing part and the flat part, and the boundary of the
flat part and the increasing part. The model is assumed to have a gap between
the left part and the middle part, and a gap between the flat part and the
right part. This algorithm is based on non-parametric MLE.
}
