% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpectralCoordinates.R
\name{getSpectralCoordinates}
\alias{getSpectralCoordinates}
\title{getSpectralCoordinates}
\usage{
getSpectralCoordinates(A, n_eigen_vectors = 0)
}
\arguments{
\item{A}{similarity matrix}

\item{n_eigen_vectors}{number of eigen vectors to save for clustering}
}
\value{
List with the following components
\itemize{
\item \code{coordinate_matrix} - matrix of spectral coordinates for each
gene
\item \code{eigen_values} - eigen values corresponding to columns of
\code{coordinate_matrix}.
}
}
\description{
Internal function to compute spectral coordinates from similarity matrix
}
