% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUMAPCoordinates.R
\name{getUMAPCoordinates}
\alias{getUMAPCoordinates}
\title{getUMAPCoordinates}
\usage{
getUMAPCoordinates(A, n_eigen_vectors = 0, n_neighbors = 15)
}
\arguments{
\item{A}{similarity matrix}

\item{n_eigen_vectors}{number of eigen vectors to save for clustering}
}
\value{
\code{coordinate_matrix} - matrix of UMAP coordinates for each gene.
}
\description{
Internal function to compute UMAP coordinates from similarity matrix
}
