% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulin_secretion-data.R
\docType{data}
\name{insulin_secretion}
\alias{insulin_secretion}
\title{GO term genes for insulin secretion GO term (GO:0030073)}
\format{
An object of class \code{character} of length 320.
}
\source{
\href{http://www.informatics.jax.org/go/term/GO:0030073}{MGD}
}
\usage{
data(insulin_secretion)
}
\description{
GO term genes for insulin secretion GO term (GO:0030073)
}
\examples{
data(insulin_secretion)
head(insulin_secretion)
}
\references{
Bult CJ, et al., The Mouse Genome Database Group. 2019.
Mouse Genome Database (MGD) 2019. Nucleic Acids Res. 2019 Jan. 8;47 (D1): D801–D806.
(\href{http://www.informatics.jax.org/go/term/GO:0030073}{MGD})
}
\keyword{datasets}
