% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performSplitUMAP.R
\name{mat_performSplitUMAP}
\alias{mat_performSplitUMAP}
\title{performSplitUMAP}
\usage{
mat_performSplitUMAP(
  potential_bait,
  exp_mat,
  n_rounds,
  round,
  alpha,
  n_neighbors,
  min_genes
)
}
\description{
Internal function for probing fishability of a set of genes to split up
potential bait set by doing clustering with k = 2.
}
