% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancreas-data.R
\docType{data}
\name{pancreas}
\alias{pancreas}
\title{scRNA seq matrix of pancreas cells}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5220 rows and 849 columns.
}
\source{
\href{https://tabula-muris.ds.czbiohub.org/}{Tabula Muris database}
}
\usage{
data(pancreas)
}
\description{
Data from Tabula Muris project.  A subset of the single cell RNA-seq data
from the pancreas to be used as an example for GeneFishing.  The rows are
genes and columns are cells.
}
\examples{
data(pancreas)
pancreas[1:4, 1:4]
}
\references{
Schaum, N. et al. (2018) Nature 562, 367–372
(\href{https://www.nature.com/articles/s41586-018-0590-4}{Nature})
}
\keyword{datasets}
