% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performSplit.R
\name{performSplit}
\alias{performSplit}
\title{performSplit}
\usage{
performSplit(potential_bait, exp_mat, n_rounds, round, alpha)
}
\description{
Internal function for probing fishability of a set of genes to split up
potential bait set by doing clustering with k = 2.
}
