% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performSplitUMAP.R
\name{performSplitUMAP}
\alias{performSplitUMAP}
\title{performSplitUMAP}
\usage{
performSplitUMAP(
  potential_bait,
  exp_mat,
  n_rounds,
  round,
  alpha,
  n_neighbors,
  min_genes,
  method
)
}
\arguments{
\item{potential_bait}{potential bait genes}

\item{exp_mat}{expression matrix}

\item{n_rounds}{number of rounds of sampling}

\item{round}{round index}

\item{alpha}{number of genes to sample}

\item{n_neighbors}{number of neighbors for UMAP}

\item{min_genes}{minimum number of genes}

\item{method}{similarity method}
}
\description{
Internal function for probing fishability of a set of genes to split up
potential bait set by doing clustering with k = 2.
}
