% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectralCoordinates.R
\name{plotSpectralCoordinates}
\alias{plotSpectralCoordinates}
\title{plotSpectralCoordinates}
\usage{
plotSpectralCoordinates(
  X,
  potential_bait,
  alpha = 5,
  umap = TRUE,
  method = c("spearman", "pearson", "cosine", "euclidean", "maximum", "manhattan",
    "canberra", "binary", "minkowski")
)
}
\arguments{
\item{X}{matrix of gene expression where the rows are genes and the
columns are samples (cells or individuals).}

\item{potential_bait}{set of genes for which to check if there is a
subset that can be used as bait.  These genes should be a subset of the
rownames of \code{X}.}

\item{alpha}{controls number of random genes that are sampled in each round
of fishing.  It will sample \code{alpha} times the number of genes in
\code{bait_genes}.  The default is 5.  The stronger the bait separates from
samples of random genes the larger \code{alpha} the algorithm can handle.}

\item{umap}{indicator of whether the computation should be done using UMAP.
The default is TRUE.  When umap = FALSE it will use spectral coordinates.}

\item{method}{input to \code{cor()} function telling what type of correlation
to use. Defaults to spearman.}

\item{k}{number of eigen vectors to use in UMAP.}
}
\value{
Plot of gene projection.
}
\description{
Function to plot the eigen decomposition of a set of potential bait genes
and a randomly sample group of additional genes. Will plot the UMAP of those
coordinates if
}
\examples{

}
