% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeFishabilityUMAP.R
\name{probeFishabilityUMAP}
\alias{probeFishabilityUMAP}
\title{probeFishabilityUMAP}
\usage{
probeFishabilityUMAP(
  exp_mat,
  potential_bait,
  n_rounds,
  alpha,
  min_tightness,
  min_genes,
  n_neighbors
)
}
\arguments{
\item{exp_mat}{matrix of gene expression where the rows are genes and the
columns are samples (cells or individuals).}

\item{n_rounds}{number of random samples to be used in assessing fishability
of a set of genes.}

\item{alpha}{controls number of random genes that are sampled in each round
of fishing.  It will sample \code{alpha} times the number of genes in
\code{bait_genes}.}

\item{min_tightness}{parameter that controls how tight the set of genes needs
to be to be considered as bait.  Tightness is measured computing the ratio of
the median distance between all potential bait genes to the distance between the
medoid of the potential bait genes and a set of randomly sampled genes. The
average of these tightness measures are taken over \code{n_rounds} random samples
of genes.}

\item{min_genes}{minimum size that the bait set can be.}

\item{potential_bait_genes}{set of genes for which to check if there is a
subset that can be used as bait.  These genes should be a subset of the
rownames of \code{exp_mat}.}
}
\description{
Function to find a set of genes that can be used as bait in geneFishing from
a larger set of potential bait genes.
}
