% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAvgDBIndexSpectral.R
\name{sce_computeAvgDBIndexSpectral}
\alias{sce_computeAvgDBIndexSpectral}
\title{computeAvgDBIndexSpectral}
\usage{
sce_computeAvgDBIndexSpectral(
  bait_genes,
  exp_mat,
  k = 2,
  alpha = 5,
  n_rounds = 50
)
}
\description{
Internal function to compute tightness of gene set when probing for fishability
}
