% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performSplitSpectral.R
\name{sce_performSplitSpectral}
\alias{sce_performSplitSpectral}
\title{performSplitSpectral}
\usage{
sce_performSplitSpectral(potential_bait, exp_mat, n_rounds, round, alpha)
}
\description{
Internal function for probing fishability of a set of genes to split up
potential bait set by doing clustering with k = 2.
}
