% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAvgDBIndexCosUMAP.R
\name{computeAvgDBIndexCosUMAP}
\alias{computeAvgDBIndexCosUMAP}
\title{computeAvgDBIndexCosUMAP}
\usage{
computeAvgDBIndexCosUMAP(
  bait_genes,
  exp_mat,
  method,
  k = 2,
  alpha = 5,
  n_rounds = 50,
  n_neighbors = 15
)
}
\description{
Internal function to compute tightness of gene set when probing for fishability
}
