% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneFishing.R
\name{geneFishing}
\alias{geneFishing}
\title{geneFishing}
\usage{
geneFishing(bait_genes, exp_mat, alpha = 5, fishing_rounds = 1000, k = 2)
}
\arguments{
\item{bait_genes}{genes to use as bait to fish out similar genes from the
\code{exp_mat}.}

\item{exp_mat}{matrix of gene expression where the rows are genes and the
columns are samples (cells or individuals).}

\item{alpha}{controls number of random genes that are sampled in each round
of fishing.  It will sample \code{alpha} times the number of genes in
\code{bait_genes}.  The default is 5.  The stronger the bait separates from
samples of random genes the larger \code{alpha} the algorithm can handle.}

\item{fishing_rounds}{number of rounds of fishing to do.  The default is 1,000.}

\item{k}{number of clusters to use in geneFishing.  The default is 2, however
if you fished out too many genes with \code{k = 2} it can help to increase to
\code{k = 3}.}
}
\value{
A data.frame with the capture frequency rate for all genes
}
\description{
Function to perform gene fishing on a gene expression matrix.
}
\examples{

}
