% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performSplitUMAP.R
\name{performSplitUMAP}
\alias{performSplitUMAP}
\title{performSplitUMAP}
\usage{
performSplitUMAP(
  potential_bait,
  exp_mat,
  n_rounds,
  round,
  alpha,
  n_neighbors,
  min_genes,
  method
)
}
\description{
Internal function for probing fishability of a set of genes to split up
potential bait set by doing clustering with k = 2.
}
