% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeFishability.R
\name{probeFishability}
\alias{probeFishability}
\title{probeFishability}
\usage{
probeFishability(
  X,
  potential_bait,
  n_rounds = 100,
  alpha = 5,
  min_tightness = 0.5,
  min_genes = 5,
  n_neighbors = 15,
  umap = TRUE,
  ncores = 2,
  method = c("spearman", "pearson", "cosine", "euclidean", "maximum", "manhattan",
    "canberra", "binary", "minkowski")
)
}
\arguments{
\item{X}{matrix of gene expression where the rows are genes and the
columns are samples (cells or individuals).}

\item{n_rounds}{number of random samples to be used in assessing fishability
of a set of genes.  The default is 100.}

\item{alpha}{controls number of random genes that are sampled in each round
of fishing.  It will sample \code{alpha} times the number of genes in
\code{bait_genes}.  The default is 5.}

\item{min_tightness}{parameter that controls how tight the set of genes needs
to be to be considered as bait.  Tightness is measured computing the ratio of
the median distance between all potential bait genes to the distance between the
medoid of the potential bait genes and a set of randomly sampled genes. The
average of these tightness measures are taken over \code{n_rounds} random samples
of genes.}

\item{min_genes}{minimum size that the bait set can be.  The default is set to
5.}

\item{umap}{indicator of whether the computation should be done using UMAP.
The default is FALSE.  When umap = TRUE the computation will take longer.
However, if you cannot find good bait with umap = FALSE, you may want to try
with umap = TRUE.}

\item{ncores}{number of cores to do parallel computation on if parallel = TRUE.}

\item{method}{input to \code{cor()} or \code{dist()} functions telling what
type of correlation to use. Cosine similarity also a possibility.
Defaults to spearman.  This must be one of "spearman", "pearson", "cosine",
"euclidean",  "maximum", "manhattan", "canberra", "binary", or "minkowski".}

\item{potential_bait_genes}{set of genes for which to check if there is a
subset that can be used as bait.  These genes should be a subset of the
rownames of \code{X}.}
}
\value{
List with the following components:
\itemize{
\item \code{best_bait} - tightest set of genes found.
\item \code{bait_sets} - list with all tight sets of genes and the corresponding
tightness metric.
\item \code{bait_info} - data.frame with the number and tightness metric
of each of the discovered bait sets. The \code{bait_index} column corresponds
to the slot in the \code{bait_sets} list (e.g. \code{bait_index} = 1 has
the metadata for \code{bait_sets[[1]]})
}
}
\description{
Function to find a set of genes that can be used as bait in geneFishing from
a larger set of potential bait genes.
}
\examples{

}
