% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralClustering.R
\name{spectralClustering}
\alias{spectralClustering}
\title{spectralClustering}
\usage{
spectralClustering(cor_mat, k)
}
\arguments{
\item{cor_mat}{correlation / similarity matrix}

\item{k}{number of clusters}
}
\description{
Internal function to do spectral clustering for \code{geneFishing} and
\code{probeFishability}.
}
