% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAPClustering.R
\name{UMAPClustering}
\alias{UMAPClustering}
\title{UMAPClustering}
\usage{
UMAPClustering(cor_mat, k, n_neighbors)
}
\arguments{
\item{cor_mat}{correlation / similarity matrix}

\item{k}{number of eigen vectors}

\item{n_neighbors}{number of neighbors for UMAP}
}
\description{
Internal function to do UMAP clustering for \code{geneFishing} and
\code{probeFishability}.
}
